modify user migration, model, password logic
