+++ models/user.js
@@ -1,55 +1,26 @@
-
-'use strict';
-
+'use strict'
 const bcrypt = require('bcryptjs')
 
 module.exports = (sequelize, DataTypes) => {
   const User = sequelize.define('User', {
     email: {
-      type: DataTypes.STRING(189),
+      type: DataTypes.STRING,
       unique: true
     },
-    password: DataTypes.STRING(189),
-    name: DataTypes.STRING(189),
-    avatar: DataTypes.STRING(189),
+    password: DataTypes.STRING,
+    name: DataTypes.STRING,
+    avatar: DataTypes.STRING,
     introduction: DataTypes.TEXT,
     account: {
-      type: DataTypes.STRING(189),
+      type: DataTypes.STRING,
       unique: true
     },
-    cover: DataTypes.STRING(189),
+    cover: DataTypes.STRING,
     role: {
-      type: DataTypes.STRING(189),
+      type: DataTypes.STRING,
       defaultValue: 'user'
     }
-  }, {
-    hooks: {
-      beforeCreate: async (User, next) => {
-        const password = User.dataValues.password
-        if (password) {
-          try {
-            const salt = await bcrypt.genSalt(10)
-            User.dataValues.password = await bcrypt.hash(password, salt)
-          }
-          catch (error) {
-            console.log(error)
-          }
-        }
-      },
-      beforeUpdate: async (User, next) => {
-        const password = User.dataValues.password
-        if (password) {
-          try {
-            const salt = await bcrypt.genSalt(10)
-            User.dataValues.password = await bcrypt.hash(password, salt)
-          }
-          catch (error) {
-            console.log(error)
-          }
-        }
-      }
-    }
-  }, {});
+  }, {})
   User.associate = function (models) {
     User.hasMany(models.Tweet, {
       foreignKey: 'UserId',
@@ -89,6 +60,6 @@ module.exports = (sequelize, DataTypes) => {
       foreignKey: 'followingId',
       as: 'Followers'
     })
-  };
-  return User;
-};
+  }
+  return User
+}
